pre2.3.0 change 86: Sprinkled SLindex_Type, SLstrlen_Type, and SLFUTURE_CONST here and there.
86. Numerous changes from int -> SLindex_Type.  Note that SLindex_Type
    is currently typedefed to be an int.  For version 3, SLindex_Type
    will become a long.  This change lays some of that groundwork.
    Similarly, I added SLFUTURE_CONST here and there.  Neither of
    these changes break binary compatibility nor affect the API.

+++ src/slassoc.c
@@ -583,7 +583,7 @@ static int
 transfer_element (SLang_Class_Type *cl, VOID_STAR dest_data,
 		  SLang_Object_Type *obj)
 {
-   unsigned int sizeof_type;
+   size_t sizeof_type;
    VOID_STAR src_data;
 
 #if USE_NEW_ANYTYPE_CODE
